+++ src/dollar/index.spec.ts
@@ -1,12 +1,11 @@
-import Dollar from '.';
-import Money from '../money';
+import Money, { Dollar } from '../money';
 
 describe('Multi currency money', () => {
   it.todo('should sum $5 + 10CHF as being $10 if exchange rate is 2:1');
 
   describe('multiplication', () => {
     it('should multiply $5 by 2 and get $10', () => {
-      const five = Money.dollar(5);
+      const five: Money = Money.dollar(5);
       const product = five.times(2);
       expect(product.equals(new Dollar(10))).toBe(true);
     });
