--- src/dollar/index.ts
@@ -1,11 +1,0 @@
-import Money from '../money';
-
-export default class Dollar extends Money {
-  constructor(amount: number) {
-    super(amount);
-  }
-
-  times(multiplier: number): Dollar {
-    return new Dollar(this.amount * multiplier);
-  }
-}
